+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.4.0] - 2018-04-
+
+### Added
+
+- Added datepicker component.
+- Added introduction.
+
+### Changed
+
+- Changed destination for script and styles. They are no longer named after version, but instead put in a folder with the current version (ex `/0.4.0/scripts/px.js`).
+
 ## [0.3.1] - 2018-04-19
 
 - Removed old version of `px-script`.